put buildDeployments in servicehistoryutils class
