+++ veteran-verification/src/main/java/gov/va/api/lighthouse/veteranverification/service/utils/ServiceHistoryUtils.java
@@ -1,5 +1,9 @@
 package gov.va.api.lighthouse.veteranverification.service.utils;
 
+import gov.va.api.lighthouse.veteranverification.api.v0.Deployment;
+import java.time.LocalDate;
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.SneakyThrows;
 import lombok.experimental.UtilityClass;
 import org.apache.commons.lang3.StringUtils;
@@ -17,7 +21,6 @@ public class ServiceHistoryUtils {
   @SneakyThrows
   public String buildBranchOfServiceString(String branchOfService, String personnelCategory) {
     String branch = null;
-
     switch (StringUtils.normalizeSpace(branchOfService.toUpperCase())) {
       case "O":
         branch = "NOAA";
@@ -43,9 +46,7 @@ public class ServiceHistoryUtils {
       default:
         throw new Exception("Invalid Branch Of Service");
     }
-
     String category;
-
     switch (StringUtils.normalizeSpace(personnelCategory)) {
       case "N":
         category = "National Guard";
@@ -58,6 +59,24 @@ public class ServiceHistoryUtils {
         category = "";
     }
     return StringUtils.normalizeSpace(String.format("%s %s", branch, category));
+  }
+
+  /**
+   * Creates a list of deployments in range of the provided start and end dates.
+   *
+   * @param deployments Super list of deployments.
+   * @param startDate Service episode startDate.
+   * @param endDate Service episode endDate.
+   * @return All deployments in range of start and end dates.
+   */
+  public List<Deployment> buildDeployments(
+      List<Deployment> deployments, LocalDate startDate, LocalDate endDate) {
+    return deployments.stream()
+        .filter(
+            deployment ->
+                isBeforeOrEqualTo(startDate, deployment.startDate())
+                    && (endDate == null || isBeforeOrEqualTo(deployment.endDate(), endDate)))
+        .collect(Collectors.toList());
   }
 
   /**
@@ -74,11 +93,14 @@ public class ServiceHistoryUtils {
         || payGradeCode.trim().isEmpty()) {
       return "unknown";
     }
-
     return StringUtils.normalizeSpace(
         String.format(
             "%s%s",
             StringUtils.normalizeSpace(payPlanCode).charAt(1),
             StringUtils.normalizeSpace(payGradeCode)));
+  }
+
+  private boolean isBeforeOrEqualTo(LocalDate dateOne, LocalDate dateTwo) {
+    return dateOne.isBefore(dateTwo) || dateOne.isEqual(dateTwo);
   }
 }
